third_party/WebKit/LayoutTests/platform/linux/http/tests/media/media-source/mediasource-duration-expected.txt

CONSOLE ERROR: Uncaught (in promise) AbortError: The play() request was interrupted by a call to pause(). https://goo.gl/LdLk22
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = The play() request was interrupted by a call to pause(). https://goo.gl/LdLk22
PASS Test seek starts on duration reduction below currentTime
PASS Test appendBuffer completes previous seek to reduced duration
PASS Test endOfStream completes previous seek to reduced duration
PASS Test setting same duration multiple times does not fire duplicate durationchange
PASS Test duration reduction below highest buffered presentation time is disallowed
Harness: the test ran to completion.
